+++ src/app/components/modal/ConfirmModal.scss
